+++ test_apps/counter_app/lib/main.app.framy.dart
@@ -8,6 +8,8 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:device_preview/device_preview.dart';
 import 'package:counter_app/main.dart';
+import 'package:framy_annotation/framy_annotation.dart';
+import 'dart:core';
 
 void main() {
   runApp(FramyApp());